Completed top3 student sorter

+++ C# Star Practices/StudentsDataManager/Objects/Lessons.cs
@@ -3,14 +3,14 @@ namespace StudentsDataManager.Objects;
 public class Lesson
 {
     public string Name { get; init; }
-    public int Score { get; set; }
+    public double Score { get; set; }
     
     public Lesson(string name)
     {
         this.Name = name;
     }
 
-    public Lesson GetMarkedLesson(int score){
+    public Lesson GetMarkedLesson(double score){
         return new Lesson(this.Name)
         {
             Score = score
